requirements.txt
mysql-connector-python
PyMuPDF  # for reading PDF files
